+++ wp-content/plugins/wp-graphql-yoast-seo-master/CHANGELOG.md
@@ -5,22 +5,61 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.0.0] - 2019-12-01
+## [4.1.0] - 2020-05-06
 
-## Changed
+### Added
+
+- Page / Tax SEO - opengraphUrl, opengraphSiteName, opengraphPublishedTime, opengraphModifiedTime - Thanks @izzygld
+- Global SEO > Schema - siteName, wordpressSiteName, siteUrl - Thanks @izzygld
+- Contributor Images added to readme
+
+## [4.0.1] - 2020-05-03
+
+### Added
+
+- Support for Yoast 14
+- Breadcrumbs for post type and taxonomies
+- canonical url for post type and taxonomies
+
+## [3.3.0] - 2020-04-01
+
+### Added
+
+Support for Yoast configuration data including:
+
+- Webmaster verification
+- Social profiles
+- Schemas
+- Breadcrumbs (global config)
+
+## [3.2.0] - 2020-03-09
+
+### Added
+
+Adding canonical url (thanks @dafky2000 )
+
+## [3.1.0] - 2020-03-09
+
+### Added
+
+WooCommerce product support
+
+## [3.0.0] - 2020-02-04
+
+### Changed
 
 - BREAKING CHANGE - image urls are now returned as `mediaItem` type.
   This applies to `twitterImage` and `opengraphImage`
 
 ## [2.1.1] - 2019-12-01
 
-## Changed
+### Changed
 
 - Update composer
 
 ## [2.1.0] - 2019-12-01
 
-## Changed
+### Changed
 
 - Fixed term data